Use networking to hold network related pkgs
Change names of unclear methods

Use net as pkg name for short
