+++ pkg/genericapiserver/genericapiserver.go
@@ -43,6 +43,7 @@ import (
 	"k8s.io/kubernetes/pkg/storage"
 	"k8s.io/kubernetes/pkg/ui"
 	"k8s.io/kubernetes/pkg/util"
+	utilnet "k8s.io/kubernetes/pkg/util/net"
 	"k8s.io/kubernetes/pkg/util/sets"
 
 	systemd "github.com/coreos/go-systemd/daemon"
@@ -212,7 +213,7 @@ type Config struct {
 	ServiceReadWriteIP net.IP
 
 	// The range of ports to be assigned to services with type=NodePort or greater
-	ServiceNodePortRange util.PortRange
+	ServiceNodePortRange utilnet.PortRange
 
 	// Used to customize default proxy dial/tls options
 	ProxyDialer          apiserver.ProxyDialerFunc
@@ -237,7 +238,7 @@ type Config struct {
 type GenericAPIServer struct {
 	// "Inputs", Copied from Config
 	ServiceClusterIPRange *net.IPNet
-	ServiceNodePortRange  util.PortRange
+	ServiceNodePortRange  utilnet.PortRange
 	cacheTimeout          time.Duration
 	MinRequestTimeout     time.Duration
 
@@ -319,7 +320,7 @@ func setDefaults(c *Config) {
 		// We should probably allow this for clouds that don't require NodePort to do load-balancing (GCE)
 		// but then that breaks the strict nestedness of ServiceType.
 		// Review post-v1
-		defaultServiceNodePortRange := util.PortRange{Base: 30000, Size: 2768}
+		defaultServiceNodePortRange := utilnet.PortRange{Base: 30000, Size: 2768}
 		c.ServiceNodePortRange = defaultServiceNodePortRange
 		glog.Infof("Node port range unspecified. Defaulting to %v.", c.ServiceNodePortRange)
 	}
@@ -454,7 +455,7 @@ func NewHandlerContainer(mux *http.ServeMux) *restful.Container {
 func (s *GenericAPIServer) init(c *Config) {
 
 	if c.ProxyDialer != nil || c.ProxyTLSClientConfig != nil {
-		s.ProxyTransport = util.SetTransportDefaults(&http.Transport{
+		s.ProxyTransport = utilnet.SetTransportDefaults(&http.Transport{
 			Dial:            c.ProxyDialer,
 			TLSClientConfig: c.ProxyTLSClientConfig,
 		})
