+++ pkg/util/net/port_split.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package util
+package net
 
 import (
 	"strings"
