+++ pkg/util/proxy/dial.go
@@ -25,14 +25,14 @@ import (
 
 	"github.com/golang/glog"
 
-	"k8s.io/kubernetes/pkg/util"
+	utilnet "k8s.io/kubernetes/pkg/util/net"
 	"k8s.io/kubernetes/third_party/golang/netutil"
 )
 
 func DialURL(url *url.URL, transport http.RoundTripper) (net.Conn, error) {
 	dialAddr := netutil.CanonicalAddr(url)
 
-	dialer, _ := util.Dialer(transport)
+	dialer, _ := utilnet.Dialer(transport)
 
 	switch url.Scheme {
 	case "http":
@@ -45,7 +45,7 @@ func DialURL(url *url.URL, transport http.RoundTripper) (net.Conn, error) {
 		var tlsConfig *tls.Config
 		var tlsConn *tls.Conn
 		var err error
-		tlsConfig, _ = util.TLSClientConfig(transport)
+		tlsConfig, _ = utilnet.TLSClientConfig(transport)
 
 		if dialer != nil {
 			// We have a dialer; use it to open the connection, then
